Clean up unused argument
For some reason get_test_results used to add 'ns' arguments to entries
from the Nameserver module. These are interpolated into the message
anyway, and no other entries are returned with raw arguments. In #1092
we accidentally broke this behavior, but since the behavior was
undocument and unintended, it's better to clean it up than to fix it.